Better line polygon generation

+++ yabause/src/yglshaderes.c
@@ -1147,7 +1147,7 @@ int YglProgramChange( YglLevel * level, int prgid )
       }
 #else
       level->prg[level->prgcurrent].maxQuad = 12*64;
-      if ((level->prg[level->prgcurrent].quads = (int *) malloc(level->prg[level->prgcurrent].maxQuad * sizeof(int))) == NULL)
+      if ((level->prg[level->prgcurrent].quads = (float *) malloc(level->prg[level->prgcurrent].maxQuad * sizeof(float))) == NULL)
          return -1;
 
       if ((level->prg[level->prgcurrent].textcoords = (float *) malloc(level->prg[level->prgcurrent].maxQuad * sizeof(float) * 2)) == NULL)
